requirements.txt
discord.py>=2.3
requests>=2.28
gspread>=5.0
google-auth>=2.0
